refactor(politics-tracker/page/people): rename to `/person` page
