+++ packages/politics-tracker/components/person/content-item.js
