+++ packages/politics-tracker/components/person/content-link.js
