+++ packages/politics-tracker/components/person/content-list.js
