+++ packages/politics-tracker/components/person/content-title.js
