+++ packages/politics-tracker/components/person/content.js
