+++ packages/politics-tracker/components/person/list-item.js
