+++ packages/politics-tracker/components/person/profile-image.js
