+++ packages/politics-tracker/components/person/section-body-election.js
