+++ packages/politics-tracker/components/person/section-body.js
