+++ packages/politics-tracker/components/person/section-list.js
