+++ packages/politics-tracker/components/person/section-toggle.js
