+++ packages/politics-tracker/graphql/query/person/get-person-basic-info.graphql
