+++ packages/politics-tracker/graphql/query/person/get-person-elections.graphql
